use charm snap instead of a pip pkg
To build on bionic python3 needs to be used. A better solution is to use
a snap since:

1) it is python3;
2) it does not matter if xenial, bionic or something else is used;
3) it is recommended by the documentation of the charm tools repo
https://github.com/juju/charm-tools#juju-charm-tools

The only requirement is that the charm snap is either already available
on the target environment or a user has sudo privileges to install it.

For automated builds it has to be passwordless sudo to avoid password
prompts.